PTRENG-4192 - Moved a few more shared functions and renamed them. Also created packages for various pieces
